refactor: fix all comments issues
