--- docs/temp/plugin-downloads.json
@@ -1,71 +1,0 @@
-{
-  "@nomiclabs/hardhat-ethers": 272306,
-  "@nomiclabs/hardhat-waffle": 259957,
-  "solidity-coverage": 184409,
-  "@nomiclabs/hardhat-etherscan": 167612,
-  "hardhat-gas-reporter": 164517,
-  "@typechain/hardhat": 142765,
-  "hardhat-deploy": 107158,
-  "hardhat-watcher": 86464,
-  "@openzeppelin/hardhat-upgrades": 71599,
-  "hardhat-contract-sizer": 69373,
-  "hardhat-deploy-ethers": 67737,
-  "@nomiclabs/hardhat-web3": 42885,
-  "hardhat-abi-exporter": 36550,
-  "@nomiclabs/hardhat-truffle5": 31390,
-  "@tenderly/hardhat-tenderly": 30748,
-  "@nomiclabs/hardhat-solhint": 27942,
-  "hardhat-tracer": 14424,
-  "@defi-wonderland/smock": 14370,
-  "hardhat-preprocessor": 13083,
-  "hardhat-dependency-compiler": 10984,
-  "hardhat-spdx-license-identifier": 10635,
-  "hardhat-log-remover": 10077,
-  "hardhat-docgen": 9089,
-  "hardhat-time-n-mine": 7035,
-  "@primitivefi/hardhat-dodoc": 6332,
-  "@nomiclabs/hardhat-ganache": 5796,
-  "hardhat-etherscan-abi": 5034,
-  "hardhat-storage-layout": 3507,
-  "@nomiclabs/hardhat-solpp": 3387,
-  "hardhat-output-validator": 3305,
-  "hardhat-deploy-tenderly": 2931,
-  "hardhat-ethernal": 2414,
-  "@nomiclabs/hardhat-vyper": 2147,
-  "uniswap-v3-deploy-plugin": 1623,
-  "@shardlabs/starknet-hardhat-plugin": 1459,
-  "hardhat-exposed": 1085,
-  "@idle-finance/hardhat-proposals-plugin": 1030,
-  "hardhat-change-network": 1026,
-  "hardhat-fund-link": 990,
-  "@eth-optimism/plugins/hardhat/compiler": 936,
-  "@eth-optimism/plugins/hardhat/ethers": 936,
-  "@mangrovedao/hardhat-test-solidity": 848,
-  "hardhat-local-networks-config-plugin": 681,
-  "xdeployer": 613,
-  "hardhat-packager": 596,
-  "hardhat-react": 582,
-  "hardhat-circom": 423,
-  "@aragon/buidler-aragon": 401,
-  "hardhat-erc1820": 393,
-  "hardhat-gas-trackooor": 290,
-  "hardhat-diamond-abi": 258,
-  "@ericxstone/hardhat-blockscout-verify": 243,
-  "@reef-defi/hardhat-reef": 211,
-  "hardhat-test-utils": 161,
-  "@nomiclabs/hardhat-truffle4": 116,
-  "@nomiclabs/hardhat-web3-legacy": 103,
-  "hardhat-laika": 103,
-  "@unlock-protocol/hardhat-plugin": 89,
-  "hardhat-hethers": 82,
-  "hardhat-awesome-cli": 70,
-  "hardhat-proxy": 54,
-  "hardhat-sourcify": 54,
-  "hardhat-forta": 48,
-  "hardhat-multibaas-plugin": 47,
-  "@muzamint/hardhat-etherspot": 26,
-  "@georacle/hardhat-georacle": 25,
-  "hardhat-network-metadata": 22,
-  "buidler-source-descriptor": 20,
-  "@controlcpluscontrolv/hardhat-yul": 20
-}
